backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg>=0.30.0
aiosqlite==0.20.0
greenlet>=3.0.0

# AI / LLM
anthropic==0.39.0
langchain==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.0

# LLM Observability & Cost Tracing
langsmith==0.1.130

# Market Data
httpx==0.27.0
yfinance>=0.2.36

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart>=0.0.6

# WebSockets (included in fastapi)
websockets==13.0

# Rate Limiting
slowapi==0.1.9

# Utilities
uuid6==2024.7.10
pydantic-settings>=2.0.0

# Error Tracking (optional — set SENTRY_DSN env var)
sentry-sdk[fastapi]>=2.0.0
